separate WelcomeScreen across files
Should be able to have functioning app with separated files, ability to connect to front end and use webplayer, needs to finish implementation of connecting to front end.
