+++ src/index.js
@@ -1,7 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import { createRoot } from "react-dom/client";
-import "./styles.css";
 import "./Home.css";
 import "./Swipescreen.css";
